fixed massive speed bottleneck for expect macros
It seems that the expect macro keeps evaluating the matcher form for
every character, which in turn causes system call to open log file
every time. To avoid this the built-in expect macro had to be
overriden so that it binds an evaluated matcher function utside the
expect macro, before expanding it.

The init-matcher function itself was highly inefficient due to
applying formatting for pattern arguments on every new character.
Also, the regex matching was done against the entire buffered string,
which instead now is limited to the last 2000 characters instead,
before filtering for printable characters only.